+++ intrinsic_results.txt
@@ -1,0 +1,62 @@
+{
+  "AWS Transcribe": {
+    "DER": 1.3791325089386766,
+    "JER": 0.4607525521791125,
+    "MS": 4.347870759753596,
+    "MSR": 0.08120614915558744,
+    "FA": 15.17102036960986,
+    "FAR": 1.089026312288889,
+    "SE": 12.107323860369608,
+    "SER": 0.20890004749420008,
+    "Duration": 62.794004106776164,
+    "Num Speakers RMSE": 1.193755140178361
+  },
+  "Diaper": {
+    "DER": 0.8486934367024902,
+    "JER": 0.6545868156919581,
+    "MS": 8.629967546012269,
+    "MSR": 0.13402798481513442,
+    "FA": 15.501325787321063,
+    "FAR": 0.5153325607343592,
+    "SE": 9.796421451942738,
+    "SER": 0.19933289115299674,
+    "Duration": 66.56400817995909,
+    "Num Speakers RMSE": 2.323085882260903
+  },
+  "NVIDIA NeMo": {
+    "DER": 0.5109314413612491,
+    "JER": 0.3478856582407132,
+    "MS": 5.051874183673468,
+    "MSR": 0.08596120560588055,
+    "FA": 11.91890128571429,
+    "FAR": 0.2748586894301481,
+    "SE": 6.51861222448979,
+    "SER": 0.1501115463252205,
+    "Duration": 66.4649612244898,
+    "Num Speakers RMSE": 1.6830487400026448
+  },
+  "Powerser": {
+    "DER": 0.3673600385976401,
+    "JER": 0.43048468292957814,
+    "MS": 4.910741367346934,
+    "MSR": 0.0793004642505947,
+    "FA": 6.4565847959183715,
+    "FAR": 0.10936106178786067,
+    "SE": 9.585724326530608,
+    "SER": 0.17869851255918476,
+    "Duration": 71.78614489795918,
+    "Num Speakers RMSE": 0.9367769320431429
+  },
+  "Pyannote": {
+    "DER": 0.3257248989814219,
+    "JER": 0.4447576566098897,
+    "MS": 4.454789202453983,
+    "MSR": 0.07327938616131374,
+    "FA": 5.909907218813908,
+    "FAR": 0.09245250133131835,
+    "SE": 8.587135644171772,
+    "SER": 0.15999301148878975,
+    "Duration": 71.98526380368097,
+    "Num Speakers RMSE": 1.0739526214544264
+  }
+}
